+++ cli/src/commands/bookmark/create.rs
@@ -17,9 +17,11 @@ use itertools::Itertools as _;
 use jj_lib::object_id::ObjectId as _;
 use jj_lib::op_store::RefTarget;
 use jj_lib::ref_name::RefNameBuf;
+use jj_lib::repo::Repo as _;
 
 use crate::cli_util::CommandHelper;
 use crate::cli_util::RevisionArg;
+use crate::cli_util::default_ignored_remote_name;
 use crate::cli_util::has_tracked_remote_bookmarks;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
@@ -40,6 +42,10 @@ pub struct BookmarkCreateArgs {
     )]
     #[arg(add = ArgValueCompleter::new(complete::revset_expression_all))]
     revision: RevisionArg,
+
+    /// Track the bookmark on all remotes
+    #[arg(long)]
+    track: bool,
 
     /// The bookmarks to create
     #[arg(required = true, value_parser = revset_util::parse_bookmark_name)]
@@ -83,28 +89,52 @@ pub async fn cmd_bookmark_create(
     }
 
     let mut tx = workspace_command.start_transaction();
-    let remote_settings = tx.settings().remote_settings()?;
-    let remote_auto_track_matchers =
-        revset_util::parse_remote_auto_track_bookmarks_map_for_new_bookmarks(ui, &remote_settings)?;
     let readonly_repo = tx.base_repo().clone();
-    for name in bookmark_names {
-        tx.repo_mut()
-            .set_local_bookmark_target(name, RefTarget::normal(target_commit.id().clone()));
-        for (remote_name, matcher) in &remote_auto_track_matchers {
-            if !matcher.is_match(name.as_str()) {
-                continue;
+    if args.track {
+        let ignored_remote = default_ignored_remote_name(readonly_repo.store());
+        for name in bookmark_names {
+            tx.repo_mut()
+                .set_local_bookmark_target(name, RefTarget::normal(target_commit.id().clone()));
+            for (remote_name, remote_view) in readonly_repo.view().remote_views() {
+                if Some(remote_name) == ignored_remote {
+                    continue;
+                }
+                let symbol = name.to_remote_symbol(remote_name);
+                if remote_view.bookmarks.contains_key(name) {
+                    writeln!(
+                        ui.warning_default(),
+                        "Tracking bookmark that exists on the remote: {symbol}"
+                    )?;
+                }
+                tx.repo_mut().track_remote_bookmark(symbol)?;
             }
-            let Some(view) = readonly_repo.view().get_remote_view(remote_name) else {
-                continue;
-            };
-            let symbol = name.to_remote_symbol(remote_name);
-            if view.bookmarks.contains_key(name) {
-                writeln!(
-                    ui.warning_default(),
-                    "Auto-tracking bookmark that exists on the remote: {symbol}"
-                )?;
+        }
+    } else {
+        let remote_settings = tx.settings().remote_settings()?;
+        let remote_auto_track_matchers =
+            revset_util::parse_remote_auto_track_bookmarks_map_for_new_bookmarks(
+                ui,
+                &remote_settings,
+            )?;
+        for name in bookmark_names {
+            tx.repo_mut()
+                .set_local_bookmark_target(name, RefTarget::normal(target_commit.id().clone()));
+            for (remote_name, matcher) in &remote_auto_track_matchers {
+                if !matcher.is_match(name.as_str()) {
+                    continue;
+                }
+                let Some(view) = readonly_repo.view().get_remote_view(remote_name) else {
+                    continue;
+                };
+                let symbol = name.to_remote_symbol(remote_name);
+                if view.bookmarks.contains_key(name) {
+                    writeln!(
+                        ui.warning_default(),
+                        "Auto-tracking bookmark that exists on the remote: {symbol}"
+                    )?;
+                }
+                tx.repo_mut().track_remote_bookmark(symbol)?;
             }
-            tx.repo_mut().track_remote_bookmark(symbol)?;
         }
     }